backend/requirements.txt
fastapi==0.127.1
uvicorn[standard]==0.38.0
sqlalchemy==2.0.45
alembic==1.17.2
psycopg2-binary==2.9.11
celery==5.6.0
redis==7.1.0
python-multipart==0.0.21
requests==2.32.5
boto3==1.42.16
python-dotenv==1.0.1
